+++ src/config/settings.py
@@ -30,7 +30,9 @@ KAKAO_REST_API_KEY = env("KAKAO_REST_API_KEY", default="")
 INSTALLED_APPS = [
     "accounts.apps.AccountsConfig",
     "books.apps.BooksConfig",
+    "knowledge.apps.KnowledgeConfig",
     "readings.apps.ReadingsConfig",
+    "reflections.apps.ReflectionsConfig",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",