+++ src/knowledge/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class KnowledgeConfig(AppConfig):
+    """Book Knowledge 도메인 설정이다."""
+
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "knowledge"