+++ src/reflections/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class ReflectionsConfig(AppConfig):
+    """Interview와 Reflection 도메인 설정이다."""
+
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "reflections"